// src/app/interfaces/api-models/fixture-data.ts
export interface FixtureData {
  apertura?: { home: string; away: string }[][];
  clausura?: { home: string; away: string }[][];
  regionalA?: { home: string; away: string }[][];
  regionalB?: { home: string; away: string }[][];
  gruposPromotionA?: { home: string; away: string }[][];
  gruposPromotionB?: { home: string; away: string }[][];
  gruposRelegation?: { home: string; away: string }[][];
  regional1?: { home: string; away: string }[][];
  regional2?: { home: string; away: string }[][];
  regional3?: { home: string; away: string }[][];
  regional4?: { home: string; away: string }[][];
  finalA?: { home: string; away: string }[][];
  finalB?: { home: string; away: string }[][];
  finalC?: { home: string; away: string }[][];
  finalD?: { home: string; away: string }[][];
}